requirements.txt
# =====================================
# CostQ Frontend + Backend 依赖
# =====================================

# === Web 框架 ===
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
slowapi>=0.1.9
sse-starlette>=2.2.1

# === 数据库 ===
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
alembic>=1.14.0

# === AWS SDK ===
boto3>=1.35.0
botocore>=1.35.0

# === GCP SDK ===
google-auth>=2.29.0

# === 认证和安全 ===
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.2.0
cryptography>=43.0.0

# === 配置管理 ===
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1

# === 邮件 ===
aiosmtplib>=3.0.0

# === 调度 ===
apscheduler>=3.10.4
pytz>=2024.1

# === 模板 ===
jinja2>=3.1.0

# === 开发工具 ===
ruff>=0.5.0
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
